Create imagetoascii.java
This is the class containing the main method. From this class the program calls Ascii(another class) . This class has the logic of the UI in which the user will drag and drop the image which he or she wants  to convert into ASCII form. Feel free to suggest any improvements in the UI.