Create Ascii.java
This class contains the logic of showing the Ascii art to the user. This class opens a new frame where the Ascii art is shown. the background is kept black and the foreground as white so as to pop up  the image . This frame also holds the "New" button which takes the user back to the window where the original image should be dropped , thus ensuring a smooth experience of the user. Feel free to contact and suggest any improvements .